requirements.txt
cryptography==42.0.5
dict2xml==1.7.5
PyYAML==6.0.1
xmltodict==0.13.0
